+++ src/Facades/IpLogger.php
@@ -12,6 +12,7 @@ use Illuminate\Support\Facades\Facade;
  * @method static self model(string $model)
  * @method static bool|Illuminate\Database\Eloquent\Model UpdateOrCreate(\Closure $attributes, \Closure $values)
  * @method static bool|Illuminate\Database\Eloquent\Model create(\Closure $values)
+ * @method static self catch(\Closure $closure)
  * 
  * @see \AmirHossein5\LaravelIpLogger\IpLogger
  **/