+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release Notes
 
+## [v1.2.1](https://github.com/amirHossein5/laravel-ip-logger/compare/v1.1.0...v1.2.1)
+## Added
+- Catching exceptions using ```catch``` method.
+
+## Fixed
+- Fixed after exception happend not working IpLogger on the rest of the calls. 
+
 ## [v1.1.0](https://github.com/amirHossein5/laravel-ip-logger/compare/v1.0.0...v1.1.0)
 laravel 9 support.